gita_chat/requirements.txt
streamlit
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-huggingface
numpy
pandas
unidecode
torch